code/monitoring-genai-apps-with-langsmith/requirements.txt
python-dotenv==1.0.0
langchain==0.1.16
langchain-openai==0.1.3
langsmith==0.1.49
langchain-community==0.0.34
beautifulsoup4==4.12.3
langchain-chroma==0.1.0